push for dayan customer list
makita ang customer list ani
