+++ lib/services/firebase_database_services.dart
@@ -1,6 +1,6 @@
 // this is a testing for the firestore capabilities hehe
 import 'package:cloud_firestore/cloud_firestore.dart';
-//import 'package:dennis_lechon_crm/models/customer.dart';
+import 'package:dennis_lechon_crm/models/customer.dart';
 
 class CustomerService {
   // collection reference
@@ -25,5 +25,28 @@ class CustomerService {
     });
   }
 
+  List<Customer> _customerListFromSnapshot(QuerySnapshot snapshot) {
+    return snapshot.docs.map((doc){
+      //print(doc.data());
+      return Customer(
+        firstName: doc['first_name'] ?? '',
+        middleName: doc['middle_name'] ?? '',
+        lastName: doc['last_name'] ?? '',
+        //celNum: doc['cel_no'] ?? '',
+        //telNum: doc['tel_no'] ?? '',
+        //adrCity: doc['address']['city'] ?? '',
+        //adrBarangay: doc['address']['barangay'] ?? '',
+        //adrZipcode: doc['address']['zipcode'] ?? '',
+        //tagColor: doc['tag']['color'] ?? '',
+        //tagIndex: doc['tag']['index'] ?? 4,
+        //tagName: doc['tag']['tagname'] ?? '',
+        //note: doc['note']
+      );
+    }).toList();
+  }
+
+  Stream<List<Customer>>? get customers {
+    return customerCollection.snapshots().map(_customerListFromSnapshot);
+  }
 
 }
